Reorganize for hierarchy, impl-hiding
