+++ downstream/__init__.py
@@ -2,4 +2,4 @@
 
 __author__ = "Matthew Andres Moreno"
 __email__ = "m.more500@gmail.com"
-__version__ = "1.0.1"
+from ._version import __version__
