+++ setup.cfg
@@ -5,7 +5,7 @@ commit = True
 tag = True
 
 [bumpver:file_patterns]
-downstream/__init__.py =
+downstream/_version.py =
   __version__ = "{pep440_version}"
 CITATION.cff =
   version: {pep440_version}
